Feat: rename dashboard page to app page
