+++ src/app/(private)/app/page.tsx
@@ -1,0 +1,5 @@
+import { AppLayout } from '@/components/app_page/AppLayout';
+
+export default function App() {
+  return <AppLayout />;
+}
