--- src/app/(private)/dashboard/page.tsx
@@ -1,5 +1,0 @@
-import { DashboardLayout } from '@/components/dashboard_page/DashboardLayout';
-
-export default function Dashboard() {
-  return <DashboardLayout />;
-}
