+++ src/components/app_page/AppLayout.tsx
@@ -75,7 +75,7 @@ const topics = [
   }
 ];
 
-export function DashboardLayout() {
+export function AppLayout() {
   return (
     <div className="min-h-screen bg-slate-50">
       <Header />
