+++ src/components/app_page/StatCard.tsx
