+++ src/components/app_page/StatsSection.tsx
